Fix user deletion and button plural

+++ packages/near-fast-auth-signer-e2e-tests/utils/createAccount.ts
@@ -19,7 +19,12 @@ export const initializeAdmin = () => {
 };
 
 export const deleteAccount = async (userUid: string) => {
-  await admin.auth().deleteUser(userUid);
+  if (userUid) {
+    const user = await admin.auth().getUser(userUid);
+    if (user.uid) {
+      await admin.auth().deleteUser(userUid);
+    }
+  }
 };
 
 export const createAccount = async ({
